example cache with spring boot stater data redis and config

+++ redis/redis-counter/src/main/java/com/example/redis/repository/CounterRepository.java
@@ -2,12 +2,9 @@ package com.example.redis.repository;
 
 import com.example.redis.entity.Counter;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CounterRepository extends JpaRepository<Counter,Integer> {
-    Counter findCounterById(Integer id);
+public interface CounterRepository extends JpaRepository<Counter, Long> {
+    Counter findCounterById(Long id);
 }
